🚚rename: [모의면접] 폴더명 변경
- [id]를 [usename]으로 변경
- chat 폴더 이동

KAN-170
