+++ src/app/(route)/(interview)/interview/[username]/[corp]/_types/corp.ts
