+++ src/app/(route)/(interview)/interview/[username]/[corp]/_utils/listVariants.ts
