+++ src/app/(route)/(interview)/interview/[username]/[corp]/chat/_types/messageType.ts
