+++ src/app/(route)/(interview)/interview/[username]/[corp]/chat/_utils/messageVariants.ts
